Bug fixing

+++ run.sh
@@ -17,8 +17,9 @@ do
       else
         schedule="${!schedule_env_var}"
       fi
-      rm /tmp/cron.tmp
+      [ ! -f /tmp/cron.tmp ] || rm /tmp/cron.tmp
       service_url=${!env_var}
+      
 cat <<EOF >> /tmp/cron.tmp
 ${schedule} curl -X POST -H "Authorization: $TUTUM_AUTH\" -H "Accept: application/json" ${service_url}start/ &> /dev/stdout"
 EOF
